Fixes #7 - add "silencing" wrapper for assertions
A generalisation of the mechanism behind "softening", this
allows third-party assertion APIs to be wrapped so that they
don't pollute TAP stream output when they fail.
