+++ test/tests.ts
@@ -1,2 +1,2 @@
-import './test-soft-assert.js';
-import './test-suite.js';
+import './assertions-test.js';
+import './ducktest-test.js';
